Move `@api` exclusion to filter stage
When we're searching for classes in a given file, we want to get the class, no matter if `@api` is used or not.
Not returning a class when it's there feel wrong.

We detect the `@api` comment, and place that on the ClassNames object. We pass that along to FileWithClass.

Then the PossiblyUnusedClassesFilter can filter classes with `@api` tag.

Also added a few test cases.